Address code review findings
- Remove pending snapshot file (.status_output_tests.rs.pending-snap)
- Split src/testing/null_db.rs (1276 lines) into:
  - src/testing/null_db/mod.rs
  - src/testing/null_db/null_database.rs
  - src/testing/null_db/capturing_store.rs
- Add insta snapshot assertion in assert_terminal_persistence
- Add bounded property-style test test_transition_invariants_property
- Remove #![cfg(unix)] gate from sighup_reload.rs
- Convert sighup_reload.rs helpers to return Results
- Remove unnecessary #[rstest] attributes from worker_orchestrator_contract.rs
- Fix British spelling "serialise" -> "serialize" in worker_orchestrator_json_shapes.rs
